Extract common function to zero Analytics report for reuse.
